requirements.txt
asgiref==3.7.2
Django==5.0.1
django-phonenumber-field==7.3.0
django-phonenumbers==1.0.1
djangorestframework==3.14.0
phonenumbers==8.13.28
pillow==10.2.0
pytz==2024.1
sqlparse==0.4.4
